refactor(tests): split simulator tests into organized modules
Split helix/simulator/tests.rs (1659 lines) into 8 focused test modules:
- basic_tests.rs: core simulator tests
- movement_tests.rs: cursor movement operations
- editing_tests.rs: text modification operations
- selection_tests.rs: selection operations
- mode_tests.rs: mode transitions
- clipboard_tests.rs: yank/paste operations
- repeat_recording_tests.rs: recording verification
- repeat_execution_tests.rs: dot command execution

All 1246 tests pass, clippy clean.
